refactor code structure in backend
